Moves mailchimp_helper.rb to lib folder
